requirements.txt
flask
flask-cors
easyocr
opencv-python
numpy
Pillow
